requirements.txt
streamlit==1.28.0
ultralytics==8.0.227
torch==2.0.1
Pillow==9.5.0
